requirements.txt
pydantic==2.10.4
pyjwt==2.6.0
python-dotenv==1.0.1
fastapi==0.115.6
python-multipart==0.0.20
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
mypy==1.14.0
uvicorn==0.34.0
pandas==1.5.3